requirements.txt
discord.py~=2.0.1
python-dotenv~=0.21.0
oauthlib~=3.2.1
requests-oauthlib==1.3.1
requests~=2.28.1